# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Clipup Project Management & Messaging API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      parameters:
        - name: spaceId
          in: query
          schema:
            type: integer
        - name: departmentId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{id}/tasks:
    get:
      operationId: getProjectTasks
      tags: [tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: assigneeId
          in: query
          schema:
            type: integer
        - name: priority
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
  /projects/{id}/stats:
    get:
      operationId: getProjectStats
      tags: [projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectStats"

  # ─── TASKS ──────────────────────────────────────────────────────────────────
...
  /goals:
    get:
      operationId: listGoals
      tags: [goals]
      parameters:
        - name: ownerId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGoalBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    get:
      operationId: getGoal
      tags: [goals]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    patch:
      operationId: updateGoal
      tags: [goals]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGoalBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    delete:
      operationId: deleteGoal
      tags: [goals]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /goals/{id}/milestones:
    get:
      operationId: getGoalMilestones
      tags: [goals]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Milestone"
    post:
      operationId: createMilestone
      tags: [goals]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMilestoneBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
  /goals/{id}/milestones/{milestoneId}:
    patch:
      operationId: updateMilestone
      tags: [goals]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: milestoneId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMilestoneBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"

  # ─── STATS ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, on_hold, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        color:
          type: string
        spaceId:
          type: ["integer", "null"]
        spaceName:
          type: ["string", "null"]
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        ownerId:
          type: integer
        ownerName:
          type: string
        startDate:
          type: ["string", "null"]
          format: date
        dueDate:
          type: ["string", "null"]
          format: date
        taskCount:
          type: integer
        completedTaskCount:
          type: integer
        progress:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, status, priority, color, ownerId, ownerName, taskCount, completedTaskCount, progress, createdAt]
    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, on_hold, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        color:
          type: string
        spaceId:
          type: ["integer", "null"]
        departmentId:
          type: ["integer", "null"]
        ownerId:
          type: integer
        startDate:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
      required: [name, ownerId]
    UpdateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, on_hold, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        color:
          type: string
        startDate:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
    ProjectStats:
      type: object
      properties:
        taskCount:
          type: integer
        completedCount:
          type: integer
        inProgressCount:
          type: integer
        todoCount:
          type: integer
        overdueCount:
          type: integer
        progress:
          type: integer
        byPriority:
          type: object
          additionalProperties:
            type: integer
        byAssignee:
          type: array
          items:
            type: object
            properties:
              userId:
                type: integer
              name:
                type: string
              count:
                type: integer
            required: [userId, name, count]
      required: [taskCount, completedCount, inProgressCount, todoCount, overdueCount, progress, byPriority, byAssignee]
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [todo, in_progress, in_review, done, cancelled]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        projectId:
          type: ["integer", "null"]
        projectName:
          type: ["string", "null"]
        parentTaskId:
          type: ["integer", "null"]
        assigneeId:
          type: ["integer", "null"]
        assigneeName:
          type: ["string", "null"]
        assigneeAvatar:
          type: ["string", "null"]
        creatorId:
          type: integer
        creatorName:
          type: string
        dueDate:
          type: ["string", "null"]
          format: date-time
        startDate:
          type: ["string", "null"]
          format: date-time
        estimatedHours:
          type: ["number", "null"]
        actualHours:
          type: ["number", "null"]
        tags:
          type: array
          items:
            type: string
        subtaskCount:
          type: integer
        completedSubtaskCount:
          type: integer
        commentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, status, priority, creatorId, creatorName, tags, subtaskCount, completedSubtaskCount, commentCount, createdAt, updatedAt]
...
    Goal:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [on_track, at_risk, off_track, completed, not_started]
        ownerId:
          type: integer
        ownerName:
          type: string
        ownerAvatar:
          type: ["string", "null"]
        targetValue:
          type: number
        currentValue:
          type: number
        unit:
          type: string
        dueDate:
          type: ["string", "null"]
          format: date-time
        progress:
          type: integer
        milestoneCount:
          type: integer
        completedMilestoneCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, status, ownerId, ownerName, targetValue, currentValue, unit, progress, milestoneCount, completedMilestoneCount, createdAt, updatedAt]
    CreateGoalBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [on_track, at_risk, off_track, completed, not_started]
        ownerId:
          type: integer
        targetValue:
          type: number
        currentValue:
          type: number
        unit:
          type: string
        dueDate:
          type: ["string", "null"]
      required: [title, ownerId, targetValue, unit]
    UpdateGoalBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [on_track, at_risk, off_track, completed, not_started]
        targetValue:
          type: number
        currentValue:
          type: number
        unit:
          type: string
        dueDate:
          type: ["string", "null"]
    Milestone:
      type: object
      properties:
        id:
          type: integer
        goalId:
          type: integer
        title:
          type: string
        targetValue:
          type: number
        isCompleted:
          type: boolean
        dueDate:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, goalId, title, targetValue, isCompleted, createdAt]
    CreateMilestoneBody:
      type: object
      properties:
        title:
          type: string
        targetValue:
          type: number
        dueDate:
          type: ["string", "null"]
      required: [title, targetValue]
    UpdateMilestoneBody:
      type: object
      properties:
        title:
          type: string
        targetValue:
          type: number
        isCompleted:
          type: boolean
        dueDate:
          type: ["string", "null"]
...