# lib/api-zod/src/generated/types/createGoalBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clipup Project Management & Messaging API
 * OpenAPI spec version: 0.2.0
 */
import type { CreateGoalBodyStatus } from "./createGoalBodyStatus";

export interface CreateGoalBody {
  title: string;
  /** @nullable */
  description?: string | null;
  status?: CreateGoalBodyStatus;
  ownerId: number;
  targetValue: number;
  currentValue?: number;
  unit: string;
  /** @nullable */
  dueDate?: string | null;
}