# lib/api-zod/src/generated/types/projectStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clipup Project Management & Messaging API
 * OpenAPI spec version: 0.2.0
 */
import type { ProjectStatsByAssigneeItem } from "./projectStatsByAssigneeItem";
import type { ProjectStatsByPriority } from "./projectStatsByPriority";

export interface ProjectStats {
  taskCount: number;
  completedCount: number;
  inProgressCount: number;
  todoCount: number;
  overdueCount: number;
  progress: number;
  byPriority: ProjectStatsByPriority;
  byAssignee: ProjectStatsByAssigneeItem[];
}